Rename collection folder to collections
