+++ services/backend/drizzle.config.ts
@@ -6,7 +6,7 @@ if(!process.env.DATABASE_URL) {
  
 export default defineConfig({
   out: './src/db/migrations',
-  schema: './src/db/tables/index.ts',
+  schema: './src/db/collections/index.ts',
   dialect: 'postgresql',
   dbCredentials: {
     url: process.env.DATABASE_URL,
