+++ services/backend/src/auth/middleware.ts
@@ -1,6 +1,6 @@
-import { Schema } from "$db/collection";
+import { Schema } from "$db/collections";
 import Elysia, { error } from "elysia";
-import type { Session, User } from "../db/collection";
+import type { Session, User } from "../db/collections";
 import { setSessionTokenCookie, validateSessionToken } from "./lucia";
 
 export const AuthService = new Elysia({ name: "Service.Auth" })
