+++ services/backend/src/db/collections/factory_areas/schema.ts
