+++ services/backend/src/db/collections/parts_to_system_models/schema.ts
