+++ services/backend/src/db/collections/readings/schema.ts
