+++ services/backend/src/db/collections/sessions/queries.ts
