+++ services/backend/src/db/collections/sessions/schema.ts
