+++ services/backend/src/db/collections/system_models/schema.ts
