+++ services/backend/src/db/collections/systems_to_factory_areas/schema.ts
