+++ services/backend/src/db/collections/systems_to_parts/schema.ts
