+++ services/backend/src/db/collections/users_to_factory_areas/schema.ts
