+++ services/backend/src/db/postgres.ts
@@ -1,6 +1,6 @@
 import { exit } from "node:process";
 import { environment } from "$config/environment";
-import { systems } from "$db/collection/systems/schema";
+import { systems } from "$db/collections/systems/schema";
 import { exponentialBackoff } from "$utils/exponentialBackoff";
 import { drizzle } from "drizzle-orm/postgres-js";
